Move template related methods to their own file
